+++ internal/provider/storage/storage_snapshot_policy_resource_test.go
@@ -41,12 +41,26 @@ func TestAccStorageSnapshotPolicyResource(t *testing.T) {
 			},
 			// Test importing a resource
 			{
-				ResourceName:  "netapp-ontap_snapshot_policy.example",
-				ImportState:   true,
-				ImportStateId: fmt.Sprintf("%s,%s,%s", "tf_acc_snapshot_policy", "tf_acc_svm", "cluster4"),
+				ResourceName:      "netapp-ontap_snapshot_policy.example",
+				ImportState:       true,
+				ImportStateId:     fmt.Sprintf("%s,%s,%s", "tf-sn-policy", "tf_acc_svm", "cluster4"),
+				ImportStateVerify: true,
+			},
+			// Create cluster scoped storage snapshot policy and read
+			{
+				Config: testAccStorageSnapshotPolicyClusterResourceConfig("tf-cluster-sn-policy", "create a cluster scoped test snapshot policy", true),
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr("netapp-ontap_snapshot_policy.example", "name", "tf_acc_snapshot_policy"),
+					resource.TestCheckResourceAttr("netapp-ontap_snapshot_policy.cluster_example", "name", "tf-cluster-sn-policy"),
+					resource.TestCheckResourceAttr("netapp-ontap_snapshot_policy.cluster_example", "comment", "create a cluster scoped test snapshot policy"),
+					resource.TestCheckResourceAttr("netapp-ontap_snapshot_policy.cluster_example", "enabled", "true"),
 				),
+			},
+			// Test importing a cluster scoped storage snapshot policy
+			{
+				ResourceName:      "netapp-ontap_snapshot_policy.cluster_example",
+				ImportState:       true,
+				ImportStateId:     fmt.Sprintf("%s,,%s", "tf-cluster-sn-policy", "cluster4"),
+				ImportStateVerify: true,
 			},
 		},
 	})
@@ -89,4 +103,42 @@ resource "netapp-ontap_snapshot_policy" "example" {
   },
   ]
 }`, host, admin, password, name, svmname, comment, enabled)
+}
+
+func testAccStorageSnapshotPolicyClusterResourceConfig(name string, comment string, enabled bool) string {
+	host := os.Getenv("TF_ACC_NETAPP_HOST")
+	admin := os.Getenv("TF_ACC_NETAPP_USER")
+	password := os.Getenv("TF_ACC_NETAPP_PASS")
+	if host == "" || admin == "" || password == "" {
+		fmt.Println("TF_ACC_NETAPP_HOST, TF_ACC_NETAPP_USER, and TF_ACC_NETAPP_PASS must be set for acceptance tests")
+		os.Exit(1)
+	}
+	return fmt.Sprintf(`
+provider "netapp-ontap" {
+ connection_profiles = [
+    {
+      name = "cluster4"
+      hostname = "%s"
+      username = "%s"
+      password = "%s"
+      validate_certs = false
+    },
+  ]
+}
+
+resource "netapp-ontap_snapshot_policy" "cluster_example" {
+  # required to know which system to interface with
+  cx_profile_name = "cluster4"
+  name = "%s"
+  comment = "%s"
+  enabled = "%t"
+  copies = [
+  {
+	count = 3
+	schedule = {
+	  name = "daily"
+	}
+  },
+  ]
+}`, host, admin, password, name, comment, enabled)
 }